+++ include/ShaderBuilder.hpp
@@ -1,7 +1,8 @@
 #include "ResourceManager.hpp"
 #include "VulkanContext.hpp"
-#include "image.hpp"
+// #include "image.hpp"
 #include <algorithm>
+#include <cstdio>
 #include <cstring>
 #include <memory>
 #include <tuple>
@@ -11,6 +12,8 @@
 #include <type_traits>
 #include <string>
 #include <vulkan/vulkan_core.h>
+#include "Binding.hpp"
+
 
 enum ShaderType
 {
@@ -18,151 +21,6 @@ enum ShaderType
     SHADER_FRAGMENT = VK_SHADER_STAGE_FRAGMENT_BIT,
     SHADER_GEOMETRY = VK_SHADER_STAGE_GEOMETRY_BIT,
 };
-
-
-template <typename Resource, VkDescriptorType ResourceType, int BindingCount, int Count>
-struct Binding
-{
-    using infoType = Resource;
-    static constexpr VkDescriptorType type() { return ResourceType; }
-    static constexpr int get_binding() { return BindingCount; }
-    static constexpr int get_descriptor_count() { return Count; }
-
-    VkDescriptorType bindingType() { return type(); }
-    int binding() { return get_binding(); }
-    // using type = Info;
-
-    Binding(VkDevice device, std::vector<VkDescriptorSet>& descriptors, Resource info[BindingCount])
-    {
-        std::memcpy(resources, info, BindingCount * sizeof(Resource));
-        for (int i = 0; i < BindingCount; i++)
-        {
-            for (int frame = 0; frame < MAX_FRAMES_IN_FLIGHT; frame++)
-            {
-                write(device, descriptors[frame], i, frame);
-            }
-        }
-    }
-
-    void write(VkDevice device, VkDescriptorSet& descriptorSet, int element, int frame)
-    {
-        VkWriteDescriptorSet descriptorWrite;
-        descriptorWrite.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-        descriptorWrite.dstSet = descriptorSet;
-        descriptorWrite.dstBinding = binding();
-        descriptorWrite.dstArrayElement = element;
-        descriptorWrite.descriptorType = type();
-        descriptorWrite.descriptorCount = 1;
-        descriptorWrite.pNext = nullptr;
-        writeResource(descriptorWrite, frame);
-
-        vkUpdateDescriptorSets(device, 1, &descriptorWrite, 0, nullptr);
-    }
-
-    virtual void writeResource(VkWriteDescriptorSet& descriptorWrite, int frame) = 0;
-
-    Resource resources[BindingCount];
-};
-
-template <typename T>
-inline constexpr bool is_binding_v =
-    std::is_base_of_v<Binding<typename T::infoType, T::type(), T::get_binding(), T::get_descriptor_count()>, T>;
-
-struct StorageImageResource
-{
-    Image image;
-
-    StorageImageResource(Image image) : image{image} {}
-};
-
-template <int binding, int bindingCount>
-struct StorageImageBinding : Binding<StorageImageResource, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, binding, bindingCount>
-{
-    StorageImageBinding(VkDevice device, std::vector<VkDescriptorSet>& descriptors, StorageImageResource info[bindingCount]) 
-    : Binding<StorageImageResource, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, binding, bindingCount>(device, descriptors, info)
-    {}
-
-    void writeResource(VkWriteDescriptorSet& descriptorWrite, int frame) override
-    {
-        
-    }
-};
-
-struct BufferBindingResource
-{
-
-};
-
-template <int binding, int bindingCount>
-struct BufferBinding : Binding<BufferBindingResource, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, binding, bindingCount>
-{
-    BufferBinding(VkDevice device, std::vector<VkDescriptorSet>& descriptors, BufferBindingResource info[bindingCount]) 
-    : Binding<BufferBindingResource, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, binding, bindingCount>(device, descriptors, info)
-    {}
-
-    void writeResource(VkWriteDescriptorSet& descriptorWrite, int frame) override
-    {
-
-    }
-};
-
-template <int Binding, int DescriptorCount>
-struct BindingSlot {
-    template <typename T>
-    struct Bind {
-        using infoType = T;
-        T info[DescriptorCount];
-        
-        // Add constructor
-        Bind(const T (&arr)[DescriptorCount]) {
-            std::copy(std::begin(arr), std::end(arr), std::begin(info));
-        }
-        
-        // For single-element initialization
-        Bind(const T& value) requires (DescriptorCount == 1) {
-            info[0] = value;
-        }
-        
-        static constexpr int get_binding() { return Binding; }
-        static constexpr int get_descriptor() { return DescriptorCount; }
-    };
-};
-
-template <typename type, int binding>
-struct Bind {
-    static_assert(is_binding_v<type>);
-    using infoType = typename type::infoType;
-    static constexpr VkDescriptorType get_type() { return type::type(); }
-    static constexpr int get_binding() { return binding; }
-};
-
-template <int binding, typename Tuple>
-struct find_binding;
-
-template <int binding, typename First, typename... Rest>
-struct find_binding<binding, std::tuple<First, Rest...>> {
-    using type = typename std::conditional_t<
-        (First::get_binding() == binding),
-        First,
-        typename find_binding<binding, std::tuple<Rest...>>::type
-    >;
-};
-
-template <int binding>
-struct find_binding<binding, std::tuple<>> {
-    using type = void;
-};
-
-// Check if a binding exists in a tuple
-template <int binding, typename Tuple>
-struct has_binding {
-    static constexpr bool value = 
-        !std::is_same_v<typename find_binding<binding, Tuple>::type, void>;
-};
-
-// Get the type of a binding in a tuple
-template <int binding, typename Tuple>
-using get_binding_type = typename find_binding<binding, Tuple>::type;
 
 template <typename... Shaders>
 struct has_duplicate_shader_types;
@@ -183,34 +41,6 @@ struct has_duplicate_shader_types<First, Rest...>
 template <typename... Shaders>
 struct all_shader_types_unique {
     static constexpr bool value = !has_duplicate_shader_types<Shaders...>::value;
-};
-
-// Combine bindings from multiple shaders
-template <typename... Shaders>
-struct CombinedBindings {
-    using AllBindings = decltype(std::tuple_cat(
-        std::declval<typename Shaders::BindingsList>()...
-    ));
-    
-    template <typename Tuple, typename Result = std::tuple<>>
-    struct filter_duplicates;
-    
-    template <typename... Result>
-    struct filter_duplicates<std::tuple<>, std::tuple<Result...>> {
-        using type = std::tuple<Result...>;
-    };
-    
-    template <typename First, typename... Rest, typename... Result>
-    struct filter_duplicates<std::tuple<First, Rest...>, std::tuple<Result...>> {
-        static constexpr int current_binding = First::get_binding();
-        using type = typename std::conditional_t<
-            has_binding<current_binding, std::tuple<Result...>>::value,
-            typename filter_duplicates<std::tuple<Rest...>, std::tuple<Result...>>::type,
-            typename filter_duplicates<std::tuple<Rest...>, std::tuple<Result..., First>>::type
-        >;
-    };
-    
-    using type = typename filter_duplicates<AllBindings>::type;
 };
 
 // Validate that matching bindings have the same type
@@ -417,6 +247,8 @@ public:
     
     ShaderPipeline(std::unique_ptr<VulkanContext>& ctx, UniqueBindingResources resources, Shaders&... shaders) : m_shaders(shaders...),
     descriptorSetLayout([&]() {
+        printf("Creating descriptor set layout\n");
+        fflush(stdout);
         std::vector<VkDescriptorSetLayoutBinding> descriptorBindings;
         std::unordered_map<int, int> bindingStages;
 
@@ -464,6 +296,8 @@ public:
         return layout;
     }()),
     descriptorPool([&](){
+        printf("Creating descriptor Pool\n");
+        fflush(stdout);
         std::unordered_map<VkDescriptorType, int> descriptorCounts;
 
         std::apply([&](auto&&... bindings) {
@@ -474,20 +308,19 @@ public:
         }, bindings);
 
         std::vector<VkDescriptorPoolSize> poolSizes;
-        poolSizes.reserve(descriptorCounts.size());
-        for (auto&& [descriptorType, descriptorCount] : descriptorCounts)
+        for (auto& [descriptorType, descriptorCount] : descriptorCounts)
         {
             VkDescriptorPoolSize poolSize;
             poolSize.type = descriptorType;
             poolSize.descriptorCount = descriptorCount;
-            poolSizes.push_back(poolSize);
+            poolSizes.push_back(std::move(poolSize));
         }
 
         VkDescriptorPoolCreateInfo createInfo;
         createInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
         createInfo.poolSizeCount = poolSizes.size();
         createInfo.pPoolSizes = poolSizes.data();
-        createInfo.maxSets = MAX_FRAMES_IN_FLIGHT;
+        createInfo.maxSets = 50;
         
         VkDescriptorPool pool;
         if (vkCreateDescriptorPool(ctx->getDevice(), &createInfo, nullptr, &pool) != VK_SUCCESS)
@@ -513,23 +346,29 @@ public:
         return std::move(descriptorSets);
     }()),
     bindings([&]() {
-
-    return std::apply([&](auto&&... resources) {
-
-        return [&]<std::size_t... Is>(std::index_sequence<Is...>) {
-            using ResultTuple = SortedBindings;
-            return ResultTuple{
-                // Ensure we're using the correct construction method
-                [&]<size_t I>(auto&& resource) {
-                    using BindingType = std::tuple_element_t<I, ResultTuple>;
-                    return BindingType{ctx->getDevice(), descriptorSets, resource.info};
-                }.template operator()<Is>(resources)...
-            };
-        }(std::make_index_sequence<sizeof...(resources)>{});
-    }, std::move(resources));
-}())
+        return std::apply([&](auto&&... resources) {
+            return [&]<std::size_t... Is>(std::index_sequence<Is...>) {
+                using ResultTuple = SortedBindings;
+                return ResultTuple{
+                    // Ensure we're using the correct construction method
+                    [&]<size_t I>(auto&& resource) {
+                        using BindingType = std::tuple_element_t<I, ResultTuple>;
+                        return BindingType{ctx->getDevice(), descriptorSets, resource.info};
+                    }.template operator()<Is>(resources)...
+                };
+            }(std::make_index_sequence<sizeof...(resources)>{});
+        }, std::move(resources));
+    }())
     {
-
+        std::apply([&](auto&&... bindings){
+            ([&](auto&& binding){
+                for (int frame = 0; frame < MAX_FRAMES_IN_FLIGHT; frame++)
+                {
+                    for (int i = 0; i < binding.get_descriptor_count(); i++)
+                    binding.write(ctx->getDevice(), descriptorSets[i], i, frame);
+                }
+            }(bindings), ...);
+        }, std::forward<SortedBindings>(bindings));
     }
     
 private:
